Drop dependency on flycheck-nimsuggest
A major mode like this should not depend on opinionated frameworks
like `flycheck`, but this package does so implicitly by depending on
`flycheck-nimsuggest`. Flymake or LSP users need not have flycheck
unnecessarily installed on their behalf.

This commit drops the declared dependency and makes it clear that it
is optional and for the user to decide. _Flymake_ support will
continue to work out of the box for all users of Emacs 26+.